+++ test/runtests.jl
@@ -14,6 +14,10 @@ end
     include("nhyp_smearing.jl")
 end
 
+@testset "Native UV smearing" begin
+    include("uv_smearing.jl")
+end
+
 @testset "Molecular dynamics driver" begin
     include("md_driver.jl")
 end